feat: Add handleMember function to MemberContext for retrieving member information
